requirements.txt
jsonschema>=4.0
pytest>=7.0
python-dotenv>=1.0
pymupdf>=1.24
pytesseract>=0.3.10
Pillow>=10.0
easyocr>=1.7.0
sqlalchemy>=2.0
pymysql>=1.1
cryptography>=42.0
httpx>=0.27
pydantic[email]>=2.0
pyjwt>=2.8
